requirements.txt
fireworks-ai
python-dotenv
requests

